+++ tests/FSharp.ApiDocs.Tests/AttributeTests.fs
@@ -1,10 +1,10 @@
 [<NUnit.Framework.TestFixture>]
-module FSharp.Formatting.MetadataFormat.AttributeTests
+module FSharp.Formatting.ApiDocs.AttributeTests
 
 open FsUnit
 open System.IO
 open NUnit.Framework
-open FSharp.Formatting.MetadataFormat
+open FSharp.Formatting.ApiDocs
 open FsUnitTyped
 open FSharp.Compiler.SourceCodeServices
 open NUnit.Framework.Internal
@@ -66,8 +66,8 @@ let info =
     "root", "http://fsprojects.github.io/FSharp.FSharp.ProjectScaffold" ]
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on Module``() =
-  let modules = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs extracts Attribute on Module``() =
+  let modules = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let modul = modules |> findModule "SingleAttributeModule"
   let attribute = modul.Attributes.Head
   attribute.Name |> shouldEqual "ObsoleteAttribute"
@@ -76,15 +76,15 @@ let ``MetadataFormat extracts Attribute on Module``() =
   attribute.NamedConstructorArguments |> shouldBeEmpty
 
 [<Test>]
-let ``MetadataFormat extracts multiple Attributes on Module``() =
-  let modules = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs extracts multiple Attributes on Module``() =
+  let modules = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let modul = modules |> findModule "MultipleAttributesModule"
   let attributes = modul.Attributes
   attributes.Length |> shouldEqual 3
 
 [<Test>]
-let ``MetadataFormat extracts Attribute with argument``() =
-  let modules = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs extracts Attribute with argument``() =
+  let modules = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let modul = modules |> findModule "SingleAttributeWithArgumentModule"
   let attribute = modul.Attributes.Head
   attribute.Name |> shouldEqual "ObsoleteAttribute"
@@ -94,8 +94,8 @@ let ``MetadataFormat extracts Attribute with argument``() =
 
 
 [<Test>]
-let ``MetadataFormat extracts Attribute with named arguments``() =
-  let modules = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs extracts Attribute with named arguments``() =
+  let modules = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let modul = modules |> findModule "SingleAttributeWithNamedArgumentsModule"
   let attribute = modul.Attributes.Head
   attribute.Name |> shouldEqual "TestAttribute"
@@ -109,24 +109,24 @@ let ``MetadataFormat extracts Attribute with named arguments``() =
   ()  
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on interface``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).TypesInfos
+let ``ApiDocs extracts Attribute on interface``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).TypesInfos
   let interface' = typeInfos |> findType "AttributeInterface"
   let attribute = interface'.Attributes.Head
   attribute.Name |> shouldEqual "TestAttribute"
   attribute.FullName |> shouldEqual "AttributeTestNamespace.TestAttribute"
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on class``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).TypesInfos
+let ``ApiDocs extracts Attribute on class``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).TypesInfos
   let class' = typeInfos |> findType "AttributeClass"
   let attribute = class'.Attributes.Head
   attribute.Name |> shouldEqual "TestAttribute"
   attribute.FullName |> shouldEqual "AttributeTestNamespace.TestAttribute"
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on value in module``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs extracts Attribute on value in module``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = typeInfos |> findModule "ContentTestModule"
   let testValue = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="testValue")
   let attribute = testValue.Attributes.Head
@@ -134,8 +134,8 @@ let ``MetadataFormat extracts Attribute on value in module``() =
   attribute.FullName |> shouldEqual "AttributeTestNamespace.TestAttribute"
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on function in module``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs extracts Attribute on function in module``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = typeInfos |> findModule "ContentTestModule"
   let testFunction = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="testFunction")
   let attribute = testFunction.Attributes.Head
@@ -143,8 +143,8 @@ let ``MetadataFormat extracts Attribute on function in module``() =
   attribute.FullName |> shouldEqual "AttributeTestNamespace.TestAttribute"
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on instance member in class``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).TypesInfos
+let ``ApiDocs extracts Attribute on instance member in class``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).TypesInfos
   let class' = typeInfos |> findType "AttributeClass"
   let testMember = class'.InstanceMembers |> Seq.find (fun v -> v.Name ="TestMember")
   let attribute = testMember.Attributes.Head
@@ -152,8 +152,8 @@ let ``MetadataFormat extracts Attribute on instance member in class``() =
   attribute.FullName |> shouldEqual "AttributeTestNamespace.TestAttribute"
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on class constructor``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).TypesInfos
+let ``ApiDocs extracts Attribute on class constructor``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).TypesInfos
   let class' = typeInfos |> findType "AttributeClass"
   let ctor = class'.Constructors |> Seq.find (fun v -> v.Details.Signature ="i:int -> AttributeClass")
   let attribute = ctor.Attributes.Head
@@ -162,8 +162,8 @@ let ``MetadataFormat extracts Attribute on class constructor``() =
   attribute.NamedConstructorArguments |>  shouldContain ("String", box "ctor")
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on static member in class``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).TypesInfos
+let ``ApiDocs extracts Attribute on static member in class``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).TypesInfos
   let class' = typeInfos |> findType "AttributeClass"
   let staticMember = class'.StaticMembers |> Seq.find (fun v -> v.Name ="TestStaticMember")
   let attribute = staticMember.Attributes.Head
@@ -171,8 +171,8 @@ let ``MetadataFormat extracts Attribute on static member in class``() =
   attribute.FullName |> shouldEqual "AttributeTestNamespace.TestAttribute"
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on union case``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).TypesInfos
+let ``ApiDocs extracts Attribute on union case``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).TypesInfos
   let union = typeInfos |> findType "AttributeUnion"
   let case = union.UnionCases |> Seq.find (fun v -> v.Name ="TestCase")
   let attribute = case.Attributes.Head
@@ -181,8 +181,8 @@ let ``MetadataFormat extracts Attribute on union case``() =
   attribute.NamedConstructorArguments |>  shouldContain ("String", box "union")
 
 [<Test>]
-let ``MetadataFormat extracts Attribute on record field``() =
-  let typeInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).TypesInfos
+let ``ApiDocs extracts Attribute on record field``() =
+  let typeInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).TypesInfos
   let union = typeInfos |> findType "AttributeRecord"
   let case = union.RecordFields |> Seq.find (fun v -> v.Name ="TestField")
   let attribute = case.Attributes.Head
@@ -191,8 +191,8 @@ let ``MetadataFormat extracts Attribute on record field``() =
   attribute.NamedConstructorArguments |>  shouldContain ("String", box "record")
 
 [<Test>]
-let ``MetadataFormat formats attribute without arguments``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute without arguments``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="noArguments")
   let attribute = value.Attributes.Head
@@ -201,8 +201,8 @@ let ``MetadataFormat formats attribute without arguments``() =
   attribute.Format() |> shouldEqual "[<Test>]"
 
 [<Test>]
-let ``MetadataFormat formats attribute with single int argument``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with single int argument``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="singleIntArgument")
   let attribute = value.Attributes.Head
@@ -211,8 +211,8 @@ let ``MetadataFormat formats attribute with single int argument``() =
   attribute.Format() |> shouldEqual "[<IntTest(1)>]"
 
 [<Test>]
-let ``MetadataFormat formats attribute with single string argument``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with single string argument``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="singleStringArgument")
   let attribute = value.Attributes.Head
@@ -221,8 +221,8 @@ let ``MetadataFormat formats attribute with single string argument``() =
   attribute.Format() |> shouldEqual """[<StringTest("test")>]"""
 
 [<Test>]
-let ``MetadataFormat formats attribute with single bool argument``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with single bool argument``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="singleBoolArgument")
   let attribute = value.Attributes.Head
@@ -232,8 +232,8 @@ let ``MetadataFormat formats attribute with single bool argument``() =
 
 
 [<Test>]
-let ``MetadataFormat formats attribute with single array argument``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with single array argument``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="singleArrayArgument")
   let attribute = value.Attributes.Head
@@ -242,8 +242,8 @@ let ``MetadataFormat formats attribute with single array argument``() =
   attribute.Format() |> shouldEqual """[<ArrayTest([|"test1"; "test2"|])>]"""
 
 [<Test>]
-let ``MetadataFormat formats attribute with multiple arguments``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with multiple arguments``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="multipleArguments")
   let attribute = value.Attributes.Head
@@ -252,8 +252,8 @@ let ``MetadataFormat formats attribute with multiple arguments``() =
   attribute.Format() |> shouldEqual """[<MultipleTest("test", 1, [|1; 2|])>]"""
 
 [<Test>]
-let ``MetadataFormat formats attribute with multiple named arguments``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with multiple named arguments``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="multipleNamedArguments")
   let attribute = value.Attributes.Head
@@ -262,8 +262,8 @@ let ``MetadataFormat formats attribute with multiple named arguments``() =
   attribute.Format() |> shouldEqual """[<Test(Int = 1, String = "test", Array = [|"1"; "2"|])>]"""
 
 [<Test>]
-let ``MetadataFormat formats attribute with name and suffix``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with name and suffix``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="singleBoolArgument")
   let attribute = value.Attributes.Head
@@ -271,8 +271,8 @@ let ``MetadataFormat formats attribute with name and suffix``() =
   attribute.FullName |> shouldEqual "AttributeTestNamespace.BoolTestAttribute"
   attribute.FormatLongForm() |> shouldEqual "[<BoolTestAttribute(true)>]"
 [<Test>]
-let ``MetadataFormat formats attribute with fullName``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with fullName``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="singleBoolArgument")
   let attribute = value.Attributes.Head
@@ -281,8 +281,8 @@ let ``MetadataFormat formats attribute with fullName``() =
   attribute.FormatFullName() |> shouldEqual "[<AttributeTestNamespace.BoolTest(true)>]"
 
 [<Test>]
-let ``MetadataFormat formats attribute with fullName and suffix``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs formats attribute with fullName and suffix``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "FormatTestModule"
   let value = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="singleBoolArgument")
   let attribute = value.Attributes.Head
@@ -291,24 +291,24 @@ let ``MetadataFormat formats attribute with fullName and suffix``() =
   attribute.FormatFullNameLongForm() |> shouldEqual "[<AttributeTestNamespace.BoolTestAttribute(true)>]"
 
 [<Test>]
-let ``MetadataFormat IsObsolete returns true on obsolete attribute``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs IsObsolete returns true on obsolete attribute``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "ObsoleteTestModule"
   let noMessage = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="noMessage")
   noMessage.IsObsolete |> shouldEqual true
   noMessage.ObsoleteMessage |> shouldEqual ""
 
 [<Test>]
-let ``MetadataFormat IsObsolete returns true on obsolete attribute and finds obsolete message``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs IsObsolete returns true on obsolete attribute and finds obsolete message``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "ObsoleteTestModule"
   let withMessage = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="withMessage")
   withMessage.IsObsolete |> shouldEqual true
   withMessage.ObsoleteMessage |> shouldEqual "obsolete"
 
 [<Test>]
-let ``MetadataFormat IsObsolete returns false on not obsolete attribute and finds no obsolete message``() =
-  let moduleInfos = MetadataFormat.GenerateReplacements([library], info, libDirs = [testBin]).ModuleInfos
+let ``ApiDocs IsObsolete returns false on not obsolete attribute and finds no obsolete message``() =
+  let moduleInfos = ApiDocs.GenerateModel([library], info, libDirs = [testBin]).ModuleInfos
   let module' = moduleInfos |> findModule "ObsoleteTestModule"
   let notObsolete = module'.ValuesAndFuncs |> Seq.find (fun v -> v.Name ="notObsolete")
   notObsolete.IsObsolete |> shouldEqual false